pyproject.toml
[project]
name = "restaurant-review"
version = "0.1.0"
description = ""
authors = [
    {name = "LucasKim4321",email = "xowls0131@naver.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pymysql (>=1.1.1,<2.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "ipython (>=9.0.2,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
